agent/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
pillow==11.3.0
mss==10.1.0
numpy==2.0.2
opencv-python==4.12.0.88
pytesseract==0.3.13
pyautogui==0.9.54
python-dotenv==1.1.0
